Refactor: Update QuizSection with Quiz_NA content
Replaces the content of `src/pages/WorkHard/QuizSection.jsx` with the
more detailed quiz implementation from the former `src/pages/WorkHard/Quiz_NA.jsx`.

- The `IntroCard` import path within the moved content was corrected
  to `../../components/IntroCard`.
- The component name was updated from `Quiz` to `QuizSection` to
  align with the filename.
- The redundant `Quiz_NA.jsx` file has been removed.
